requirements.txt
cachetools==3.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
rsa==4.0
six==1.12.0
tqdm==4.32.1
uritemplate==3.0.0
